nextcloud/requirements.txt
# Core FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Nextcloud Integration
webdavclient3==3.14.6
caldav==1.3.9
icalendar==5.0.11
httpx==0.25.2

# Configuration & Validation
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-dateutil==2.8.2

# Templates
jinja2==3.1.2

# Utils
cryptography==41.0.7
